+++ SF3Lib/Models/Structs/X013/StatusEffect.cs
@@ -50,142 +50,142 @@ namespace SF3.Models.Structs.X013 {
 
         [TableViewModelColumn(addressField: nameof(_luck0ChanceAddr), displayOrder: 0, displayName: "Luck 0 Chance%")]
         [BulkCopy]
-        public int Luck0Chance {
+        public byte Luck0Chance {
             get => Data.GetUInt8(_luck0ChanceAddr);
-            set => Data.SetUInt8(_luck0ChanceAddr, (byte) value);
+            set => Data.SetUInt8(_luck0ChanceAddr, value);
         }
 
         [TableViewModelColumn(addressField: nameof(_luck0UnknownAddr), displayOrder: 0.5f, displayName: "Luck 0 Unknown")]
         [BulkCopy]
-        public int Luck0Unknown {
+        public byte Luck0Unknown {
             get => Data.GetUInt8(_luck0UnknownAddr);
-            set => Data.SetUInt8(_luck0UnknownAddr, (byte) value);
+            set => Data.SetUInt8(_luck0UnknownAddr, value);
         }
 
         [TableViewModelColumn(addressField: nameof(_luck1ChanceAddr), displayOrder: 1, displayName: "Luck 1 Chance%")]
         [BulkCopy]
-        public int Luck1Chance {
+        public byte Luck1Chance {
             get => Data.GetUInt8(_luck1ChanceAddr);
-            set => Data.SetUInt8(_luck1ChanceAddr, (byte) value);
+            set => Data.SetUInt8(_luck1ChanceAddr, value);
         }
 
         [TableViewModelColumn(addressField: nameof(_luck1UnknownAddr), displayOrder: 1.5f, displayName: "Luck 1 Unknown")]
         [BulkCopy]
-        public int Luck1Unknown {
+        public byte Luck1Unknown {
             get => Data.GetUInt8(_luck1UnknownAddr);
-            set => Data.SetUInt8(_luck1UnknownAddr, (byte) value);
+            set => Data.SetUInt8(_luck1UnknownAddr, value);
         }
 
         [TableViewModelColumn(addressField: nameof(_luck2ChanceAddr), displayOrder: 2, displayName: "Luck 2 Chance%")]
         [BulkCopy]
-        public int Luck2Chance {
+        public byte Luck2Chance {
             get => Data.GetUInt8(_luck2ChanceAddr);
-            set => Data.SetUInt8(_luck2ChanceAddr, (byte) value);
+            set => Data.SetUInt8(_luck2ChanceAddr, value);
         }
 
         [TableViewModelColumn(addressField: nameof(_luck2UnknownAddr), displayOrder: 2.5f, displayName: "Luck 2 Unknown")]
         [BulkCopy]
-        public int Luck2Unknown {
+        public byte Luck2Unknown {
             get => Data.GetUInt8(_luck2UnknownAddr);
-            set => Data.SetUInt8(_luck2UnknownAddr, (byte) value);
+            set => Data.SetUInt8(_luck2UnknownAddr, value);
         }
 
         [TableViewModelColumn(addressField: nameof(_luck3ChanceAddr), displayOrder: 3, displayName: "Luck 3 Chance%")]
         [BulkCopy]
-        public int Luck3Chance {
+        public byte Luck3Chance {
             get => Data.GetUInt8(_luck3ChanceAddr);
-            set => Data.SetUInt8(_luck3ChanceAddr, (byte) value);
+            set => Data.SetUInt8(_luck3ChanceAddr, value);
         }
 
         [TableViewModelColumn(addressField: nameof(_luck3UnknownAddr), displayOrder: 3.5f, displayName: "Luck 3 Unknown")]
         [BulkCopy]
-        public int Luck3Unknown {
+        public byte Luck3Unknown {
             get => Data.GetUInt8(_luck3UnknownAddr);
-            set => Data.SetUInt8(_luck3UnknownAddr, (byte) value);
+            set => Data.SetUInt8(_luck3UnknownAddr, value);
         }
 
         [TableViewModelColumn(addressField: nameof(_luck4ChanceAddr), displayOrder: 4, displayName: "Luck 4 Chance%")]
         [BulkCopy]
-        public int Luck4Chance {
+        public byte Luck4Chance {
             get => Data.GetUInt8(_luck4ChanceAddr);
-            set => Data.SetUInt8(_luck4ChanceAddr, (byte) value);
+            set => Data.SetUInt8(_luck4ChanceAddr, value);
         }
 
         [TableViewModelColumn(addressField: nameof(_luck4UnknownAddr), displayOrder: 4.5f, displayName: "Luck 4 Unknown")]
         [BulkCopy]
-        public int Luck4Unknown {
+        public byte Luck4Unknown {
             get => Data.GetUInt8(_luck4UnknownAddr);
-            set => Data.SetUInt8(_luck4UnknownAddr, (byte) value);
+            set => Data.SetUInt8(_luck4UnknownAddr, value);
         }
 
         [TableViewModelColumn(addressField: nameof(_luck5ChanceAddr), displayOrder: 5, displayName: "Luck 5 Chance%")]
         [BulkCopy]
-        public int Luck5Chance {
+        public byte Luck5Chance {
             get => Data.GetUInt8(_luck5ChanceAddr);
-            set => Data.SetUInt8(_luck5ChanceAddr, (byte) value);
+            set => Data.SetUInt8(_luck5ChanceAddr, value);
         }
 
         [TableViewModelColumn(addressField: nameof(_luck5UnknownAddr), displayOrder: 5.5f, displayName: "Luck 5 Unknown")]
         [BulkCopy]
-        public int Luck5Unknown {
+        public byte Luck5Unknown {
             get => Data.GetUInt8(_luck5UnknownAddr);
-            set => Data.SetUInt8(_luck5UnknownAddr, (byte) value);
+            set => Data.SetUInt8(_luck5UnknownAddr, value);
         }
 
         [TableViewModelColumn(addressField: nameof(_luck6ChanceAddr), displayOrder: 6, displayName: "Luck 6 Chance%")]
         [BulkCopy]
-        public int Luck6Chance {
+        public byte Luck6Chance {
             get => Data.GetUInt8(_luck6ChanceAddr);
-            set => Data.SetUInt8(_luck6ChanceAddr, (byte) value);
+            set => Data.SetUInt8(_luck6ChanceAddr, value);
         }
 
         [TableViewModelColumn(addressField: nameof(_luck6UnknownAddr), displayOrder: 6.5f, displayName: "Luck 6 Unknown")]
         [BulkCopy]
-        public int Luck6Unknown {
+        public byte Luck6Unknown {
             get => Data.GetUInt8(_luck6UnknownAddr);
-            set => Data.SetUInt8(_luck6UnknownAddr, (byte) value);
+            set => Data.SetUInt8(_luck6UnknownAddr, value);
         }
 
         [TableViewModelColumn(addressField: nameof(_luck7ChanceAddr), displayOrder: 7, displayName: "Luck 7 Chance%")]
         [BulkCopy]
-        public int Luck7Chance {
+        public byte Luck7Chance {
             get => Data.GetUInt8(_luck7ChanceAddr);
-            set => Data.SetUInt8(_luck7ChanceAddr, (byte) value);
+            set => Data.SetUInt8(_luck7ChanceAddr, value);
         }
 
         [TableViewModelColumn(addressField: nameof(_luck7UnknownAddr), displayOrder: 7.5f, displayName: "Luck 7 Unknown")]
         [BulkCopy]
-        public int Luck7Unknown {
+        public byte Luck7Unknown {
             get => Data.GetUInt8(_luck7UnknownAddr);
-            set => Data.SetUInt8(_luck7UnknownAddr, (byte) value);
+            set => Data.SetUInt8(_luck7UnknownAddr, value);
         }
 
         [TableViewModelColumn(addressField: nameof(_luck8ChanceAddr), displayOrder: 8, displayName: "Luck 8 Chance%")]
         [BulkCopy]
-        public int Luck8Chance {
+        public byte Luck8Chance {
             get => Data.GetUInt8(_luck8ChanceAddr);
-            set => Data.SetUInt8(_luck8ChanceAddr, (byte) value);
+            set => Data.SetUInt8(_luck8ChanceAddr, value);
         }
 
         [TableViewModelColumn(addressField: nameof(_luck8UnknownAddr), displayOrder: 8.5f, displayName: "Luck 8 Unknown")]
         [BulkCopy]
-        public int Luck8Unknown {
+        public byte Luck8Unknown {
             get => Data.GetUInt8(_luck8UnknownAddr);
-            set => Data.SetUInt8(_luck8UnknownAddr, (byte) value);
+            set => Data.SetUInt8(_luck8UnknownAddr, value);
         }
 
         [TableViewModelColumn(addressField: nameof(_luck9ChanceAddr), displayOrder: 9, displayName: "Luck 9 Chance%")]
         [BulkCopy]
-        public int Luck9Chance {
+        public byte Luck9Chance {
             get => Data.GetUInt8(_luck9ChanceAddr);
-            set => Data.SetUInt8(_luck9ChanceAddr, (byte) value);
+            set => Data.SetUInt8(_luck9ChanceAddr, value);
         }
 
         [TableViewModelColumn(addressField: nameof(_luck9UnknownAddr), displayOrder: 9.5f, displayName: "Luck 9 Unknown")]
         [BulkCopy]
-        public int Luck9Unknown {
+        public byte Luck9Unknown {
             get => Data.GetUInt8(_luck9UnknownAddr);
-            set => Data.SetUInt8(_luck9UnknownAddr, (byte) value);
+            set => Data.SetUInt8(_luck9UnknownAddr, value);
         }
     }
 }
